// src/lib/api/generated/model/serviceGuideReadinessSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */
import type { ServiceGuideMissingFieldSchema } from './serviceGuideMissingFieldSchema';

export interface ServiceGuideReadinessSchema {
  display_state: string;
  document_official: boolean;
  document_readiness: string;
  missing_fields: ServiceGuideMissingFieldSchema[];
  official: boolean;
  pending_notice: string;
  state: string;
}